refactor(persistence): co-locate worker-control + cleanup-schedule-state with jobs-worker
