+++ assistant/src/__tests__/config-watcher-skill-reseed.test.ts
@@ -82,7 +82,7 @@ mock.module("../persistence/embeddings/embedding-backend.js", () => ({
   clearEmbeddingBackendCache: () => {},
 }));
 
-mock.module("../memory/cleanup-schedule-state.js", () => ({
+mock.module("../persistence/cleanup-schedule-state.js", () => ({
   resetCleanupScheduleThrottle: () => {},
 }));
