+++ assistant/src/persistence/__tests__/worker-control.test.ts
