+++ assistant/src/persistence/cleanup-schedule-state.ts
