+++ assistant/src/persistence/jobs-worker.ts
@@ -8,16 +8,15 @@ import {
   diskPressureBackgroundSkipLogFields,
   shouldLogDiskPressureBackgroundSkip,
 } from "../daemon/disk-pressure-background-gate.js";
-import {
-  getLastScheduledCleanupEnqueueMs,
-  markScheduledCleanupEnqueued,
-} from "../memory/cleanup-schedule-state.js";
 import { sweepOrphanMemoryRetrospectiveConversations } from "../memory/memory-retrospective-startup-cleanup.js";
 import { countBufferLines } from "../memory/v2/consolidation-job.js";
-import { spawnMemoryWorkerProcess } from "../memory/worker-control.js";
 import { getLogger } from "../util/logger.js";
 import { getWorkspaceDir } from "../util/platform.js";
 import { getMemoryCheckpoint, setMemoryCheckpoint } from "./checkpoints.js";
+import {
+  getLastScheduledCleanupEnqueueMs,
+  markScheduledCleanupEnqueued,
+} from "./cleanup-schedule-state.js";
 import { maybeRunDbMaintenance } from "./db-maintenance.js";
 import {
   EmbeddingBillingBlockError,
@@ -49,6 +48,7 @@ import {
   resetRunningJobsToPending,
   SLOW_LLM_JOB_TYPES,
 } from "./jobs-store.js";
+import { spawnMemoryWorkerProcess } from "./worker-control.js";
 
 const log = getLogger("memory-jobs-worker");
